Move popover settings to `MenuController`
